refactor: use two type files
one is exposed to the user of the library
